requirements.txt
pytest-playwright
playwright
allure-pytest
